+++ api/src/main/java/org/apache/cloudstack/context/ResponseMessageResolver.java
@@ -30,10 +30,8 @@ import java.util.Map;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import org.apache.cloudstack.api.ApiErrorCode;
 import org.apache.cloudstack.api.Identity;
 import org.apache.cloudstack.api.InternalIdentity;
-import org.apache.cloudstack.api.ServerApiException;
 import org.apache.cloudstack.api.response.ExceptionResponse;
 import org.apache.commons.collections.MapUtils;
 import org.apache.commons.lang3.ObjectUtils;
@@ -41,8 +39,6 @@ import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import com.cloud.exception.InvalidParameterValueException;
-import com.cloud.exception.PermissionDeniedException;
 import com.cloud.utils.PropertiesUtil;
 import com.cloud.utils.Ternary;
 import com.cloud.utils.exception.CloudRuntimeException;
@@ -369,37 +365,5 @@ public class ResponseMessageResolver {
         Map<String, String> stringMap = getStringMap(combinedMetadata);
         response.setErrorText(expand(template,  stringMap));
         response.setErrorMetadata(stringMap);
-    }
-
-    public static InvalidParameterValueException invalidParameterValueException(String errorKey, Map<String, Object> metadata) {
-        return new InvalidParameterValueException(errorKey, metadata);
-    }
-
-    public static InvalidParameterValueException invalidParameterValueException(String errorKey) {
-        return invalidParameterValueException(errorKey, Collections.emptyMap());
-    }
-
-    public static PermissionDeniedException permissionDeniedException(String errorKey, Map<String, Object> metadata) {
-        return new PermissionDeniedException(errorKey, metadata);
-    }
-
-    public static PermissionDeniedException permissionDeniedException(String errorKey) {
-        return permissionDeniedException(errorKey, Collections.emptyMap());
-    }
-
-    public static ServerApiException serverApiException(ApiErrorCode errorCode, String errorKey) {
-        return new ServerApiException(errorCode, errorKey, Collections.emptyMap());
-    }
-
-    public static CloudRuntimeException cloudRuntimeException(String errorKey, Map<String, Object> metadata) {
-        return new CloudRuntimeException(resolve(errorKey, metadata));
-    }
-
-    public static CloudRuntimeException cloudRuntimeException(String errorKey) {
-        return new CloudRuntimeException(resolve(errorKey, Collections.emptyMap()));
-    }
-
-    public static CloudRuntimeException cloudRuntimeException(String errorKey, Map<String, Object> metadata, Throwable th) {
-        return new CloudRuntimeException(resolve(errorKey, metadata), th);
     }
 }
